+++ src/pages/index.ts
@@ -6,13 +6,13 @@ import SearchPage from "./Search/Search";
 import SettingPage from "./Setting/Setting";
 import SearchResultPage from "./SearchResult/SearchResult";
 import PotDetailPage from "./PotDetail/PotDetail";
-import PotMain from "./Pots/PotMain/PotMain";
+import PotMain from "./Pots/PotMain";
 import AllPotPage from "./Pots/pages/AllPot/AllPotPage";
 import AppliedPotPage from "./Pots/pages/AppliedPot/AppliedPot";
 import MadePotPage from "./Pots/pages/MadePot/MadePot"; 
 import SignUpPage from "./SignUp/SignUp";
 import CreatePotPage from "./CreatePot/CreatePot";
-import MyPotDetailMainPage from "./MyPotDetail/MyPotDetailMain";
+import MyPotDetailMainPage from "./MyPotDetail/MyPotDetail";
 import MyPotStatusPage from "./MyPotDetail/pages/MyPotStatus/MyPotStatus";
 import MyPotCalendarPage from "./MyPotDetail/pages/MyPotCalendar/MyPotCalendar";
 import EditPostPage from "./EditPost/EditPost";
